+++ zcash_primitives/src/transaction/builder.rs
@@ -544,7 +544,7 @@ mod testing {
 
 #[cfg(test)]
 mod tests {
-    use ff::{Field, PrimeField};
+    use ff::Field;
     use rand_core::OsRng;
 
     use crate::{
@@ -552,13 +552,13 @@ mod tests {
         legacy::TransparentAddress,
         memo::MemoBytes,
         merkle_tree::{CommitmentTree, IncrementalWitness},
-        sapling::{Node, Rseed},
+        sapling::Rseed,
         transaction::components::{
             amount::{Amount, DEFAULT_FEE},
             sapling::builder::{self as build_s},
             transparent::builder::{self as build_t},
         },
-        zip32::{ExtendedFullViewingKey, ExtendedSpendingKey},
+        zip32::ExtendedSpendingKey,
     };
 
     use super::{Builder, Error};
@@ -583,9 +583,9 @@ mod tests {
     #[test]
     fn fails_on_negative_output() {
         let extsk = ExtendedSpendingKey::master(&[]);
-        let extfvk = ExtendedFullViewingKey::from(&extsk);
-        let ovk = extfvk.fvk.ovk;
-        let to = extfvk.default_address().1;
+        let dfvk = extsk.to_diversifiable_full_viewing_key();
+        let ovk = dfvk.fvk().ovk;
+        let to = dfvk.default_address().1;
 
         let sapling_activation_height = TEST_NETWORK
             .activation_height(NetworkUpgrade::Sapling)
@@ -665,15 +665,15 @@ mod tests {
     #[test]
     fn binding_sig_present_if_shielded_spend() {
         let extsk = ExtendedSpendingKey::master(&[]);
-        let extfvk = ExtendedFullViewingKey::from(&extsk);
-        let to = extfvk.default_address().1;
+        let dfvk = extsk.to_diversifiable_full_viewing_key();
+        let to = dfvk.default_address().1;
 
         let mut rng = OsRng;
 
         let note1 = to
             .create_note(50000, Rseed::BeforeZip212(jubjub::Fr::random(&mut rng)))
             .unwrap();
-        let cmu1 = Node::new(note1.cmu().to_repr());
+        let cmu1 = note1.commitment();
         let mut tree = CommitmentTree::empty();
         tree.append(cmu1).unwrap();
         let witness1 = IncrementalWitness::from_tree(&tree);
@@ -738,9 +738,9 @@ mod tests {
             );
         }
 
-        let extfvk = ExtendedFullViewingKey::from(&extsk);
-        let ovk = Some(extfvk.fvk.ovk);
-        let to = extfvk.default_address().1;
+        let dfvk = extsk.to_diversifiable_full_viewing_key();
+        let ovk = Some(dfvk.fvk().ovk);
+        let to = dfvk.default_address().1;
 
         // Fail if there is only a Sapling output
         // 0.0005 z-ZEC out, 0.00001 t-ZEC fee
@@ -783,7 +783,7 @@ mod tests {
         let note1 = to
             .create_note(50999, Rseed::BeforeZip212(jubjub::Fr::random(&mut rng)))
             .unwrap();
-        let cmu1 = Node::new(note1.cmu().to_repr());
+        let cmu1 = note1.commitment();
         let mut tree = CommitmentTree::empty();
         tree.append(cmu1).unwrap();
         let mut witness1 = IncrementalWitness::from_tree(&tree);
@@ -823,7 +823,7 @@ mod tests {
         let note2 = to
             .create_note(1, Rseed::BeforeZip212(jubjub::Fr::random(&mut rng)))
             .unwrap();
-        let cmu2 = Node::new(note2.cmu().to_repr());
+        let cmu2 = note2.commitment();
         tree.append(cmu2).unwrap();
         witness1.append(cmu2).unwrap();
         let witness2 = IncrementalWitness::from_tree(&tree);